Improve sidebar subreddit list sorting behavior.
It turns out Backbone fires the 'sort' event after adding new models to a
collection. This allows us to reorder the views in the DOM in one batch
rather than having to insert elements into specific offsets in the DOM,
which breaks when 'add' events are out of order from the collection.